Allow blank lines and prepend spaces
Text bbox is calculated from the real drawing area.
However, this approach has these problems:

1. Cannot preserve the spaces that users put at the beginning of lines.
2. The size of the left side bearing of the first character differs
   for each glyph, so when Text elements are arranged vertically,
   they become uneven.

To solve these problems, I decided to generate bbox that includes the
leading spaces and left-side bearing. As a result, even when there are
no spaces at the beginning of a line, it generates a slightly wider
bbox to the left than before.